переносит рендеринг фильмов из main.js
